+++ collection-system/event/scale-event.h
@@ -23,8 +23,8 @@ class FSEvent: public Event {
 private:
   osm_pid_t pid;
   int inode;
-  size_t bytes_read;
-  size_t bytes_written;
+  long bytes_read;
+  long bytes_written;
 
   std::string event;
   std::string event_time;
@@ -37,7 +37,7 @@ private:
 
 public:
   FSEvent(const std::string &serialized_event);
-  FSEvent(osm_pid_t pid, int inode, size_t bytes_read, size_t bytes_written,
+  FSEvent(osm_pid_t pid, int inode, long bytes_read, long bytes_written,
       std::string event, std::string event_time, std::string cluster_name,
       std::string fs_name, std::string path, std::string dst_path,
       std::string mode, std::string version_hash);
